shared auth state to the components to speed up the ui rendering speed
