+++ frontend/src/App.jsx
@@ -1,5 +1,7 @@
-import { Suspense, lazy } from 'react'
+import { Suspense, lazy, useState, useEffect } from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import Cookies from 'universal-cookie'
+import axios from 'axios'
 
 /* Auth Components */
 const AuthRoute = lazy(() => import("./components/auth/auth_route"))
@@ -18,9 +20,34 @@ const PageNotFound = lazy(() => import('./components/pages/PageNotFound'))
 import './App.css'
 
 function App() {
+  const [isAuth, setAuth] = useState(false)
+
+  useEffect(() => {
+    /* Check authentication of the app */
+    const cookies = new Cookies()
+    const token = cookies.get("token")
+    if (token !== undefined) {
+      axios.post("http://localhost:8000/api/get-decoded-token", {
+        token: token,
+      }).then((response) => {
+        if (response.data.auth === true) {
+          setAuth(true)
+        }
+        else {
+          setAuth(false)
+        }
+      }).catch(err => {
+        console.log(err)
+      })
+    }
+    else {
+      setAuth(false)
+    }
+  }, [])
+
   return (
     <div id="body">
-      <NavigationBar />
+      <NavigationBar isAuth={isAuth} />
       <main id="content">
         <BrowserRouter>
           <Suspense fallback="Loading...">
@@ -32,7 +59,7 @@ function App() {
               <Route path="*" element={<PageNotFound />}></Route>
 
               {/* Auth Route */}
-              <Route element={<AuthRoute />}>
+              <Route element={<AuthRoute isAuth={isAuth}/>}>
                 <Route path="dashboard" element={<Dashboard />}></Route>
               </Route>
             </Routes>
